Use Sliding sync tables for bulk fetching max event pos in rooms
